pyproject.toml
[project]
name = "appetit"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12.0"
dependencies = [
    "alembic>=1.16.4",
    "argon2-cffi>=25.1.0",
    "asyncpg>=0.30.0",
    "authx>=1.4.3",
    "fastapi>=0.116.1",
    "gunicorn>=23.0.0",
    "psycopg>=3.2.9",
    "pydantic[email]>=2.11.7",
    "uvicorn>=0.35.0",
]
